capture/requirements.txt
picamera==1.13
ffmpeg-python==0.1.17
